requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
astroid==2.15.6
async-timeout==4.0.2
attrs==23.1.0
black==23.7.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
dill==0.3.7
frozenlist==1.4.0
ghp-import==2.1.0
griffe==0.32.3
idna==3.4
isort==5.12.0
Jinja2==3.1.2
lazy-object-proxy==1.9.0
Markdown==3.4.4
MarkupSafe==2.1.3
mccabe==0.7.0
mergedeep==1.3.4
mkdocs==1.5.2
mkdocs-autorefs==0.5.0
mkdocs-material==9.1.21
mkdocs-material-extensions==1.1.1
mkdocstrings==0.22.0
mkdocstrings-python==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
openai==0.27.8
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
Pygments==2.16.1
pylint==2.17.5
pymdown-extensions==10.1
PyQt6==6.5.2
PyQt6-Qt6==6.5.2
PyQt6-sip==13.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
pyyaml_env_tag==0.1
regex==2023.8.8
requests==2.31.0
six==1.16.0
tomlkit==0.12.1
tqdm==4.65.0
urllib3==2.0.4
watchdog==3.0.0
wrapt==1.15.0
yarl==1.9.2
